+++ packages/ionic-react/src/schematics/init/lib/add-dependencies.ts
@@ -3,7 +3,6 @@ import { addDepsToPackageJson } from '@nrwl/workspace';
 import {
   ioniconsVersion,
   ionicReactVersion,
-  nxtendVersion,
   testingLibraryJestDomVersion,
   testingLibraryUserEventVersion,
 } from '../../../utils/versions';
@@ -15,7 +14,6 @@ export function addDependencies(): Rule {
       ionicons: ioniconsVersion,
     },
     {
-      '@nxtend/ionic-react': nxtendVersion,
       '@testing-library/user-event': testingLibraryUserEventVersion,
       '@testing-library/jest-dom': testingLibraryJestDomVersion,
     }